[REF] split python file into 3 class;
