+++ stock_picking_reorder_lines/models/__init__.py
@@ -19,4 +19,6 @@
 #
 #
 
-from . import models
+from . import sale_order
+from . import stock_move
+from . import stock_picking
