+++ stock_picking_reorder_lines/models/stock_move.py
@@ -19,4 +19,16 @@
 #
 #
 
-from . import models
+from openerp.osv import orm, fields
+
+
+class stock_move(orm.Model):
+    _inherit = 'stock.move'
+    _columns = {
+        'sequence': fields.integer('Sequence',
+                                   help="Gives the sequence of this line when "
+                                        "displaying the picking."),
+    }
+    _order = 'date_expected desc, sequence, id'
+    _defaults = {'sequence': 10,
+                 }
